Move DotExpandingXContentParser to server
DotExpandingXContentParser is a parser that's very specific to document parsing, based on the assumption
that what gets parsed is never loaded into a map as it can expose duplicate keys. That means that we want
to make sure it does not get misused in other scenarios, hence this commit moves it out of the xcontent
library into server as a package private class.
